refactor(web-client): move sign function to session-xrpl/algorand service.
