+++ web-client/src/app/state/session-algorand.service.ts
@@ -10,8 +10,15 @@ import {
 } from 'src/app/services/algosdk.utils';
 import { EnclaveService } from 'src/app/services/enclave/index';
 import { SessionService } from 'src/app/state/session.service';
+import { withLoggedExchange } from 'src/app/utils/console.helpers';
 import { panic } from 'src/app/utils/errors/panic';
-import { TransactionSigned, TransactionToSign } from 'src/schema/actions';
+import { never } from 'src/helpers/helpers';
+import {
+  SignTransaction,
+  SignTransactionResult,
+  TransactionSigned,
+  TransactionToSign,
+} from 'src/schema/actions';
 import { SessionQuery } from './session.query';
 import { SessionStore } from './session.store';
 
@@ -104,6 +111,46 @@ export class SessionAlgorandService {
   }
 
   /**
+   * Sign a transaction using the active session's wallet.
+   *
+   * This takes care of wrapping {@link SignTransaction}
+   * and unwrapping {@link SignTransactionResult}.
+   *
+   * @see EnclaveService#signTransaction
+   */
+  async signTransaction(
+    transaction_to_sign: TransactionToSign
+  ): Promise<TransactionSigned> {
+    const { wallet, pin } = this.sessionQuery.assumeActiveSession();
+
+    const signRequest: SignTransaction = {
+      auth_pin: pin,
+      wallet_id: wallet.wallet_id,
+      transaction_to_sign,
+    };
+    const signResult: SignTransactionResult = await withLoggedExchange(
+      'SessionService: EnclaveService.signTransaction:',
+      async () => await this.enclaveService.signTransaction(signRequest),
+      signRequest
+    );
+
+    if ('Signed' in signResult) {
+      return signResult.Signed;
+    } else if ('InvalidAuth' in signResult) {
+      this.sessionStore.setError({ signResult });
+      throw panic('SessionService.signTransaction: invalid auth', signResult);
+    } else if ('Failed' in signResult) {
+      this.sessionStore.setError({ signResult });
+      throw panic(
+        `SessionService.signTransaction failed: ${signResult.Failed}`,
+        signResult
+      );
+    } else {
+      throw never(signResult);
+    }
+  }
+
+  /**
    * Helper: Sign, submit, and confirm the given transaction.
    */
   protected async sendTransaction(
@@ -112,9 +159,7 @@ export class SessionAlgorandService {
     const unsigned: TransactionToSign = {
       AlgorandTransaction: { transaction_bytes: transaction.bytesToSign() },
     };
-    const signed: TransactionSigned = await this.sessionService.signTransaction(
-      unsigned
-    );
+    const signed: TransactionSigned = await this.signTransaction(unsigned);
 
     if ('AlgorandTransactionSigned' in signed) {
       console.log('SessionAlgorandService.sendTransaction:', { signed });
