+++ web-client/src/app/state/session.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@angular/core';
 import { EnclaveService } from 'src/app/services/enclave/index';
 import { MessagingService } from 'src/app/services/messaging.service';
 import { SearchService } from 'src/app/services/search.service';
+import { SessionXrplService } from 'src/app/state/session-xrpl.service';
 import { SessionQuery } from 'src/app/state/session.query';
 import { withLoggedExchange } from 'src/app/utils/console.helpers';
 import { panic } from 'src/app/utils/errors/panic';
@@ -17,10 +18,6 @@ import {
   OpenWalletResult,
   SaveOnfidoCheck,
   SaveOnfidoCheckResult,
-  SignTransaction,
-  SignTransactionResult,
-  TransactionSigned,
-  TransactionToSign,
 } from 'src/schema/actions';
 import { SessionStore } from './session.store';
 
@@ -111,53 +108,6 @@ export class SessionService {
       throw new Error(result.Failed);
     } else {
       throw never(result);
-    }
-  }
-
-  /**
-   * Sign a transaction using the active session's wallet.
-   *
-   * This takes care of wrapping {@link SignTransaction}
-   * and unwrapping {@link SignTransactionResult}.
-   *
-   * @see EnclaveService#signTransaction
-   */
-  async signTransaction(
-    transaction_to_sign: TransactionToSign,
-    wallet_id?: string,
-    account_pin?: string
-  ): Promise<TransactionSigned> {
-    const { wallet, pin } = this.sessionQuery.assumeActiveSession();
-    const active_wallet_id = wallet.wallet_id;
-
-    const wallet_id_tx = wallet_id ? wallet_id : active_wallet_id;
-
-    const pin_tx = account_pin ? account_pin : pin;
-
-    const signRequest: SignTransaction = {
-      auth_pin: pin_tx,
-      wallet_id: wallet_id_tx,
-      transaction_to_sign,
-    };
-
-    const signResult: SignTransactionResult = await withLoggedExchange(
-      'SessionService: EnclaveService.signTransaction:',
-      async () => await this.enclaveService.signTransaction(signRequest),
-      signRequest
-    );
-    if ('Signed' in signResult) {
-      return signResult.Signed;
-    } else if ('InvalidAuth' in signResult) {
-      this.sessionStore.setError({ signResult });
-      throw panic('SessionService.signTransaction: invalid auth', signResult);
-    } else if ('Failed' in signResult) {
-      this.sessionStore.setError({ signResult });
-      throw panic(
-        `SessionService.signTransaction failed: ${signResult.Failed}`,
-        signResult
-      );
-    } else {
-      throw never(signResult);
     }
   }
